requirements.txt
Django==1.11
django-bootstrap3==11.1.0
django-tinymce==2.8.0
Pillow==6.2.1
psycopg2==2.7.7
pytz==2019.3
